+++ aesara/scan/basic.py
@@ -8,9 +8,8 @@ from aesara.compile import SharedVariable
 from aesara.compile.function.pfunc import construct_pfunc_ins_and_outs
 from aesara.configdefaults import config
 from aesara.graph.basic import Constant, Variable, clone_replace, graph_inputs
-from aesara.graph.fg import MissingInputError
 from aesara.graph.op import get_test_value
-from aesara.graph.utils import TestValueError
+from aesara.graph.utils import MissingInputError, TestValueError
 from aesara.scan import utils
 from aesara.scan.op import Scan, ScanInfo
 from aesara.scan.utils import safe_new, traverse
